Added a Plouffe's Series for base-10.

+++ Efficient Infinte Series/Double factorial/DoubleFactorial_series_1.py
@@ -16,8 +16,8 @@ pi: Decimal = Decimal(3.141592653589793238462643383279502884197169399375) # math
 #         else: result *= i
 
 k: int = 0
-approximation: float = 0
-previous: float = 0
+approximation: Decimal = 0
+previous: Decimal = 0
 
 while True:
 
@@ -34,7 +34,7 @@ while True:
             else: print(f"{i - 2} accurate decimal place(s)")
             break
 
-    deviation: float = approximation - previous
+    deviation: Decimal = approximation - previous
 
     if deviation == 0: 
         print("Negligible deviation (terminating the program)\n")
