+++ Efficient Infinte Series/Ramanujan-Sato Series/ramanujan_sato_series.py
@@ -8,7 +8,7 @@ pi: Decimal = Decimal(3.141592653589793238462643383279502884197169399375) # math
 
 k: int = 0
 approximationInverted: Decimal = 0
-previous: float = 0
+previous: Decimal = 0
 
 while True:
 
@@ -24,7 +24,7 @@ while True:
             else: print(f"{i - 2} accurate decimal place(s)")
             break
 
-    deviation: float = approximation - previous
+    deviation: Decimal = approximation - previous
 
     if deviation == 0: 
         print("Negligible deviation (terminating the program)\n")
